pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.74.0",
    "arxiv>=2.3.1",
    "mcp>=1.21.2",
    "nest-asyncio>=1.6.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "typing>=3.10.0.0",
    "uv>=0.9.10",
]
